api change: more parameters to domain validation
instead of just passing nid and port, we now pass an array of
parameters, which will usually be a array version of the node object.

hooks can now validate against a client and such.

now we are certain that client A cannot create a site on a different
port than client B for the same domain (but client A can)

see #570980