API: drop the redundant hosting_task_queue table
the table was reproducing the status field from hosting_task and was
making it difficult to reliably manipulate the status of a task.

all instances of $node->queued should now be replaced by
$node->task_status == HOSTING_TASK_QUEUED

all uses of the timestamp in the hosting_task_queue table should instead
use the task node changed field

queries that used the hosting_task_queue table for listing pending queue
should instead look directly in the hosting_task table and look for
HOSTING_TASK_QUEUED status, be sure to join with the node table to avoid
duplicate entries due to revisions

we also introduce a HOSTING_TASK_PROCESSING status for tasks since the
queue "locking" (ie. not starting two tasks at a time) was relying on
the duplicate information