Fix security faux pas related to the node access permissions.
In the most general case of the hosting default client being 1, and the client owning the site node being that client, it would not assign any permissions to that user.

the node access system would then grant view access to everyone.

this patch stops confusing the hosting_default_client and the admin client (client 1), and ensures that the permissions are correctly assigned.