Made the sites/ servers listing pages become primary links, from new installs and update.php.
Also discovered a regression, whereby the initial creation of the server node is not setting the 'available' flag to true, even if explicitly set.

this in turn is causing the installer to fail because the necessary class hasn't been instantiated.

I've reverted the check on whether a service is available before loading it's class.